frontend/requirements.txt
# Core dependencies
streamlit==1.31.1
requests==2.31.0
python-dotenv==1.0.1

# Vector database and embeddings
qdrant-client==1.7.0
sentence-transformers==2.5.1

# Database
psycopg2-binary==2.9.9

# Storage
boto3==1.34.34
botocore==1.34.34

# LangGraph and LLM dependencies
langgraph==0.5.0
langchain-core==0.3.66
langchain-groq==0.3.4
langchain-community==0.3.26
duckduckgo_search==8.0.4